Rearrange folder structure.
