+++ data-pipeline/benches/span_concentrator_bench.rs
@@ -7,7 +7,7 @@ use std::{
 
 use criterion::{criterion_group, Criterion};
 use data_pipeline::span_concentrator::SpanConcentrator;
-use datadog_trace_utils::span_v04::Span;
+use datadog_trace_utils::span::v04::Span;
 
 fn get_bucket_start(now: SystemTime, n: u64) -> i64 {
     let start = now.duration_since(time::UNIX_EPOCH).unwrap() + Duration::from_secs(10 * n);
