+++ data-pipeline/src/span_concentrator/aggregation.rs
@@ -4,7 +4,7 @@
 //! This includes the aggregation key to group spans together and the computation of stats from a
 //! span.
 use datadog_trace_protobuf::pb;
-use datadog_trace_utils::span_v04::{trace_utils, Span};
+use datadog_trace_utils::span::v04::{trace_utils, Span};
 use std::borrow::Borrow;
 use std::borrow::Cow;
 use std::collections::HashMap;
