+++ data-pipeline/src/stats_exporter.rs
@@ -187,7 +187,7 @@ pub fn stats_url_from_agent_url(agent_url: &str) -> anyhow::Result<hyper::Uri> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use datadog_trace_utils::span_v04::{trace_utils, Span};
+    use datadog_trace_utils::span::v04::{trace_utils, Span};
     use httpmock::prelude::*;
     use httpmock::MockServer;
     use time::Duration;
