+++ trace-utils/src/span/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2023-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod v04;
